Refactor location of packages
